Implemented Non CRUDs in logic

+++ HTML5-CSS_Course_Backend_Logic/Classes/TableLogic.cs
@@ -1,7 +1,8 @@
+using System.ComponentModel.DataAnnotations;
 using HTML5_CSS_Course_Backend_Models;
 using HTML5_CSS_Course_Backend_Repository;
 
-namespace HTML5_CSS_Course_Backend_Logic.Classes
+namespace HTML5_CSS_Course_Backend_Logic
 {
     public class TableLogic : ITableLogic
     {
@@ -14,42 +15,74 @@ namespace HTML5_CSS_Course_Backend_Logic.Classes
 
         public void Create(Table item)
         {
+            if (ReadAll().Contains(item))
+            {
+                throw new ArgumentException("The item you are trying to add already exists");
+            }
+            Validator.ValidateObject(item, new ValidationContext(item), true);
             repo.Create(item);
         }
 
         public void Delete(long id)
         {
+            if (ReadAll().FirstOrDefault(t => t.id == id) == null) throw new NullReferenceException("The item you are trying to delete doesn't exist");
             repo.Delete(id);
+        }
+        public void DeleteByName(string name)
+        {
+            var item = GetByName(name);
+            if (item == null) throw new NullReferenceException();
+            repo.Delete(item.id);
         }
 
         public void Disable(string name)
         {
-            throw new NotImplementedException();
+            GetByName(name)!.isEnabled = true;
         }
 
         public void Enable(string name)
         {
-            throw new NotImplementedException();
+            GetByName(name)!.isEnabled = false;
         }
 
-        public IEnumerable<Table> Free(string start, string stop)
+        public IEnumerable<Table> Free(DateTime start, DateTime stop)
         {
-            throw new NotImplementedException();
+            return ReadAll()
+                   .Where(t => t.isEnabled
+                   && t.reservations!.Where(t => t.CompareDates(start, stop)).Count() == 0)
+                   .AsEnumerable();
         }
 
-        public IEnumerable<Reservation> Get(string start, string stop)
+        public Table GetByName(string name)
         {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<Table> GetByName(string name)
-        {
-            throw new NotImplementedException();
+            Table? result;
+            try
+            {
+                result = ((TableRepository)repo).ReadByName(name);
+            }
+            catch (Exception e)
+            {
+                throw new QuerryFailedException(e);
+            }
+            if (result == null) throw new NullReferenceException();
+            return result;
         }
 
         public IEnumerable<Table> GetByState(bool state)
         {
-            throw new NotImplementedException();
+            IEnumerable<Table> result;
+            try
+            {
+                result = ReadAll()
+                        .Where(t => t.isEnabled == state)
+                        .AsEnumerable();
+            }
+            catch (Exception e)
+            {
+                throw new QuerryFailedException(e);
+            }
+            if (result == null || result.Count() == 0) throw new NullReferenceException();
+            return result;
         }
 
         public Table? Read(long id)
@@ -64,6 +97,7 @@ namespace HTML5_CSS_Course_Backend_Logic.Classes
 
         public void Update(Table item)
         {
+            Validator.ValidateObject(item, new ValidationContext(item), true);
             repo.Update(item);
         }
     }
